Extracting CommandUsageTracker
